+++ test/integration/automate/default_test.rb
@@ -1,7 +1,7 @@
-# InSpec test for recipe servicenow-ritm::automate
+# InSpec test for recipe servicenow-task::automate
 
 # script exists
-describe file('/tmp/kitchen/cache/servicenow-ritm.rb') do
+describe file('/tmp/kitchen/cache/servicenow-task.rb') do
   it { should exist }
   it { should be_executable }
   its('content') { should match /^automate_url = \"https:\/\/inez.bottlebru.sh\"$/ }
@@ -13,10 +13,10 @@ end
 
 # add crontab entry for cron[knife ec backup]
 describe crontab do
-  its('commands') { should include '/tmp/kitchen/cache/servicenow-ritm.rb' }
+  its('commands') { should include '/tmp/kitchen/cache/servicenow-task.rb' }
 end
 
-describe crontab.commands('/tmp/kitchen/cache/servicenow-ritm.rb') do
+describe crontab.commands('/tmp/kitchen/cache/servicenow-task.rb') do
   its('minutes') { should cmp '*/10' }
   its('hours') { should cmp '*' }
   its('days') { should cmp '*' }