+++ test/integration/client/default_test.rb
@@ -1,7 +1,7 @@
-# InSpec test for recipe servicenow-ritm::client
+# InSpec test for recipe servicenow-task::client
 
 describe json('/tmp/kitchen/nodes/client-centos-7.json') do
-  its(['override','servicenow','ritm']) { should eq '12345' }
+  its(['override','servicenow','task']) { should eq '12345' }
   its(['override','servicenow','source_url']) { should eq 'https://source_url' }
 end