+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# servicenow-ritm CHANGELOG
+# servicenow-task CHANGELOG
 
-This file is used to list changes made in each version of the servicenow-ritm cookbook.
+This file is used to list changes made in each version of the servicenow-task cookbook.
 
 # 0.1.0
 
 Initial release.
+
+# 0.2.0
+
+Renamed to servicenow-task